Refactoring
moving classes into subpackage form
