+++ app/src/main/java/ca/klapstein/baudit/Activities/AccountCreateActivity.java
@@ -1,6 +1,7 @@
-package ca.klapstein.baudit;
+package ca.klapstein.baudit.Activities;
 
 import android.support.annotation.Nullable;
+import ca.klapstein.baudit.Data.User;
 
 public class AccountCreateActivity {
     @Nullable
