+++ app/src/main/java/ca/klapstein/baudit/Activities/LogoutActivity.java
@@ -1,6 +1,7 @@
-package ca.klapstein.baudit;
+package ca.klapstein.baudit.Activities;
 
 import android.support.annotation.NonNull;
+import ca.klapstein.baudit.Data.User;
 
 public class LogoutActivity {
     @NonNull
