+++ app/src/main/java/ca/klapstein/baudit/Activities/ModifyProblemDialog.java
@@ -1,0 +1,7 @@
+package ca.klapstein.baudit.Activities;
+
+import ca.klapstein.baudit.Data.Problem;
+
+public class ModifyProblemDialog {
+    private Problem problem;
+}
