+++ app/src/main/java/ca/klapstein/baudit/Activities/ModifyRecordDialog.java
@@ -1,0 +1,7 @@
+package ca.klapstein.baudit.Activities;
+
+import ca.klapstein.baudit.Data.Record;
+
+public class ModifyRecordDialog {
+    private Record record;
+}
