+++ app/src/main/java/ca/klapstein/baudit/Activities/PatientListActivity.java
@@ -1,0 +1,9 @@
+package ca.klapstein.baudit.Activities;
+
+import ca.klapstein.baudit.Data.PatientTreeSet;
+import ca.klapstein.baudit.Adapters.PatientAdapter;
+
+public class PatientListActivity {
+    private PatientTreeSet patientTreeSet;
+    private PatientAdapter patientAdapter;
+}
