+++ app/src/main/java/ca/klapstein/baudit/Activities/ProblemListActivity.java
@@ -1,0 +1,9 @@
+package ca.klapstein.baudit.Activities;
+
+import ca.klapstein.baudit.Data.ProblemTreeSet;
+import ca.klapstein.baudit.Adapters.ProblemAdapter;
+
+public class ProblemListActivity {
+    private ProblemTreeSet problemTreeSet;
+    private ProblemAdapter problemAdapter;
+}
