+++ app/src/main/java/ca/klapstein/baudit/Activities/RecordListActivity.java
@@ -1,0 +1,9 @@
+package ca.klapstein.baudit.Activities;
+
+import ca.klapstein.baudit.Data.RecordTreeSet;
+import ca.klapstein.baudit.Adapters.RecordAdapter;
+
+public class RecordListActivity {
+    private RecordTreeSet recordTreeSet;
+    private RecordAdapter recordAdapter;
+}
