+++ app/src/main/java/ca/klapstein/baudit/Activities/ViewAccountDialog.java
@@ -1,0 +1,7 @@
+package ca.klapstein.baudit.Activities;
+
+import ca.klapstein.baudit.Data.User;
+
+public class ViewAccountDialog {
+    private User user;
+}
