--- app/src/main/java/ca/klapstein/baudit/BauditDataBaseManager.java
@@ -1,4 +1,0 @@
-package ca.klapstein.baudit;
-
-class BauditDataBaseManager {
-}
