--- app/src/main/java/ca/klapstein/baudit/BauditPreferencesManager.java
@@ -1,4 +1,0 @@
-package ca.klapstein.baudit;
-
-class BauditPreferencesManager {
-}
