+++ app/src/main/java/ca/klapstein/baudit/Managers/BauditPreferencesManager.java
@@ -1,0 +1,4 @@
+package ca.klapstein.baudit.Managers;
+
+class BauditPreferencesManager {
+}
