+++ app/src/main/java/ca/klapstein/baudit/Managers/BauditRemoteManager.java
@@ -1,4 +1,4 @@
-package ca.klapstein.baudit;
+package ca.klapstein.baudit.Managers;
 
 public class BauditRemoteManager {
 }
