--- app/src/main/java/ca/klapstein/baudit/ModifyProblemDialog.java
@@ -1,5 +1,0 @@
-package ca.klapstein.baudit;
-
-public class ModifyProblemDialog {
-    private Problem problem;
-}
