--- app/src/main/java/ca/klapstein/baudit/ModifyRecordDialog.java
@@ -1,5 +1,0 @@
-package ca.klapstein.baudit;
-
-public class ModifyRecordDialog {
-    private Record record;
-}
