--- app/src/main/java/ca/klapstein/baudit/PatientListActivity.java
@@ -1,6 +1,0 @@
-package ca.klapstein.baudit;
-
-public class PatientListActivity {
-    private PatientTreeSet patientTreeSet;
-    private PatientAdapter patientAdapter;
-}
