--- app/src/main/java/ca/klapstein/baudit/ProblemListActivity.java
@@ -1,6 +1,0 @@
-package ca.klapstein.baudit;
-
-public class ProblemListActivity {
-    private ProblemTreeSet problemTreeSet;
-    private ProblemAdapter problemAdapter;
-}
