--- app/src/main/java/ca/klapstein/baudit/RecordListActivity.java
@@ -1,6 +1,0 @@
-package ca.klapstein.baudit;
-
-public class RecordListActivity {
-    private RecordTreeSet recordTreeSet;
-    private RecordAdapter recordAdapter;
-}
