--- app/src/main/java/ca/klapstein/baudit/ViewAccountDialog.java
@@ -1,5 +1,0 @@
-package ca.klapstein.baudit;
-
-public class ViewAccountDialog {
-    private User user;
-}
